Conflicts:
	src/restricted/applications/indoors/indoors.pro

+++ core/graphs/undirectedGraph.cpp
@@ -11,7 +11,7 @@ void UndirectedGraph::addDGEdge(int from, int to)
 {
     dst.push_back(to);
     next.push_back(first[from]);
-    first[from] = next.size() - 1;
+    first[from] = (int)next.size() - 1;
 }
 
 void UndirectedGraph::addEdge(int from, int to)
@@ -65,7 +65,7 @@ std::vector<int> UndirectedGraph::getConnectedComponent(int v)
 
 void UndirectedGraph::intersect(const std::vector<int> &a, std::vector<int> &b)
 {
-    int N = a.size(), ptrA = 0, ptrB = 0, M = b.size(), ptrRes = 0;
+    int N = (int)a.size(), ptrA = 0, ptrB = 0, M = b.size(), ptrRes = 0;
     for (; ptrA < N && ptrB < M;)
     {
         if (a[ptrA] == b[ptrB])
